Added CountryVideo

+++ project/TIGO_videocms/trunk/grails-app/services/com/tigo/videocms/BotRInterfaceService.groovy
@@ -79,7 +79,9 @@ class BotRInterfaceService {
 	}
 
 	def deleteVideo(Video video) {
-		api.delete(video.getVideoKey())
+		if(video != null && video.getVideoKey() != null) {
+			api.delete(video.getVideoKey())
+		}
 	}
 	
 	// Helper Methods
